Removed potential to sleep in split_2M function. Added locking within ept_violation handler.
Changed way of allocating a page within the split_2M function so we shouldn't sleep (e.g. shouldn't need to  pull something from the okmm cache anymore). Added locking to the ept_violation handler: this is very coarse at the moment; we can be much more precise (and it may make sense to move to RCU based locking).